Add e2e test setup
